feat: add products and search products

+++ src/services/productsService.js
@@ -1,4 +1,5 @@
 const pool = require('../db/pool');
+const createError = require('http-errors');
 
 async function getAllProducts() {
   const { rows } = await pool.query(`
@@ -14,9 +15,10 @@ async function getAllProducts() {
     JOIN vendor_products ON products.id = vendor_products.product_id
     GROUP BY products.name, categories.name, brands.name, products.stock_qty
 `);
+
   return rows;
 }
-async function getProduct({ name, brand, category }) {
+async function getProduct(searchParam) {
   const { rows } = await pool.query(
     `
     SELECT 
@@ -29,11 +31,16 @@ async function getProduct({ name, brand, category }) {
     JOIN categories     ON products.category_id = categories.id
     JOIN brands         ON products.brand_id = brands.id
     JOIN vendor_products ON products.id = vendor_products.product_id
-    WHERE products.name=$1  OR brands.name=$2 OR categories.name=$3
+    WHERE products.name=$1  OR brands.name=$1 OR categories.name=$1
     GROUP BY products.name, categories.name, brands.name, products.stock_qty
 `,
-    [name, brand, category],
+    [searchParam],
   );
+
+  if (rows.length === 0) {
+    throw createError(404, 'Product not found');
+  }
+
   return rows;
 }
 
@@ -55,26 +62,11 @@ async function getProductDetails(id) {
 `,
     [id],
   );
-  return rows;
-}
 
-async function addProduct({ name, category_id, brand_id, vendor_id, unit_cost }) {
-  const result = await pool.query(
-    `INSERT INTO products (name, category_id, brand_id, stock_qty)
-         VALUES ($1, $2, $3, 0)
-         RETURNING *`,
-    [name, category_id, brand_id],
-  );
-
-  const product = result.rows[0];
-
-  await pool.query(
-    `INSERT INTO vendor_products (product_id, vendor_id, unit_cost)
-         VALUES ($1, $2, $3)`,
-    [product.id, vendor_id, unit_cost],
-  );
-
-  return product;
+  if (rows.length === 0) {
+    throw createError(404, 'Failed to get Product details');
+  }
+  return rows[0];
 }
 
 async function updateProduct(id, { name, category_id, brand_id, vendor_id, unit_cost }) {
@@ -88,6 +80,9 @@ async function updateProduct(id, { name, category_id, brand_id, vendor_id, unit_
     [name, category_id, brand_id, id],
   );
 
+  if (result.rows.length === 0) {
+    throw createError(404, 'Update Failed');
+  }
   // update the vendor cost if provided
   if (vendor_id && unit_cost) {
     await pool.query(
@@ -102,13 +97,34 @@ async function updateProduct(id, { name, category_id, brand_id, vendor_id, unit_
 }
 
 async function deleteProduct(id) {
+  const product = await pool.query(`SELECT id FROM products WHERE id = $1`, [id]); //check if the product exists first
+  if (product.rows.length === 0) {
+    throw createError(404, 'Product not found');
+  }
+  //remove foreign keys constrain
   await pool.query(`DELETE FROM vendor_products WHERE product_id = $1`, [id]);
 
   await pool.query(`DELETE FROM stock_movements WHERE product_id = $1`, [id]);
 
   const result = await pool.query(`DELETE FROM products WHERE id = $1 RETURNING *`, [id]);
 
+  if (result.rows.length === 0) {
+    throw createError(404, 'Delete Failed');
+  }
   return result.rows[0];
+}
+
+async function addProduct({ name, category_id, brand_id }) {
+  const result = await pool.query(
+    `INSERT INTO products (name, category_id, brand_id, stock_qty)
+         VALUES ($1, $2, $3, 0)
+         RETURNING *`,
+    [name, category_id, brand_id],
+  );
+
+  const product = result.rows[0];
+
+  return product;
 }
 
 module.exports = {
